flask_backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
tld>=0.13
